# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.3.0",
    "description": "Go-native Rebecca panel API for dashboard, bot, and third-party client developers.\n\nAuthentication: the dashboard uses an opaque server-side session stored in the HttpOnly `rebecca_session` cookie. Session permissions are loaded from the database on every request, so permission changes take effect without logging out. Bots and external integrations should use personal `rk_...` API keys as `Authorization: Bearer <api_key>`. Legacy JWT creation at `POST /admin/token` and `POST /api/admin/token` remains available only for older Marzban-compatible clients.\n\nDo not automate username/password login for bots. API keys can be revoked independently and inherit the owner admin permissions, service scope, and limits.\n\nMutation behavior: user, service, node, config, and host mutations are transactional where applicable. Runtime changes are synchronized to nodes through gRPC and recorded for recovery."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/token": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create legacy admin JWT",
        "description": "Authenticate a database-backed admin and issue a legacy JWT bearer token for older Marzban-compatible clients. The Rebecca dashboard does not call this endpoint; it uses the opaque session flow under /api/auth instead. New bots and integrations should use personal rk_ API keys rather than storing an admin password.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Token"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdminToken"
      }
    },
    "/api/admin/token": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create legacy admin JWT",
        "description": "Authenticate a database-backed admin and issue a legacy JWT bearer token for older Marzban-compatible clients. The Rebecca dashboard does not call this endpoint; it uses the opaque session flow under /api/auth instead. New bots and integrations should use personal rk_ API keys rather than storing an admin password.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Token"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdminToken"
      }
    },
    "/internal/admin/validate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Validate admin token",
        "description": "Internal validation endpoint for admin API key and dashboard JWT authentication.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postInternalAdminValidate"
      }
    },
    "/api/admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get current admin",
        "description": "Return the authenticated admin context, role, permissions, limits, and service assignments.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getAdmin"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create admin",
        "description": "Create a new admin if the current admin can manage admins.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdmin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Rebecca admin API key or legacy JWT",
        "description": "Use a personal rk_ API key for bots and third-party integrations. Legacy JWTs from /admin/token remain accepted for older Marzban-compatible clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rebecca_session",
        "description": "Opaque HttpOnly dashboard session cookie issued by POST /api/auth/login. It is managed by the browser and is not readable by dashboard JavaScript."
      }
    },
...
    "requestBodies": {
      "JSON": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Object"
            },
            "example": {
              "name": "example",
              "status": "active",
              "settings": {}
            }
          }
        },
        "description": "Generic JSON request body. See each operation description for required business fields and side effects."
      },
      "LoginForm": {
        "required": true,
        "content": {
          "application/x-www-form-urlencoded": {
            "schema": {
              "type": "object",
              "required": [
                "username",
                "password"
              ],
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "format": "password"
                }
              }
            }
          },
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "username",
                "password"
              ],
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "format": "password"
                }
              }
            }
          }
        },
        "description": "Database-backed admin username and password. JSON is used by the session dashboard endpoint; form encoding remains accepted by the legacy JWT compatibility endpoints."
      },
...
    },
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible Rebecca API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            },
            "example": {
              "id": 1,
              "status": "ok",
              "detail": "Operation completed"
            }
          }
        }
      },
      "Token": {
        "description": "Legacy JWT access token for older Marzban-compatible clients. New integrations should use personal rk_ API keys.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                },
                "token_type": {
                  "type": "string",
                  "default": "bearer"
                }
              }
            }
          }
        }
      },
      "Error": {
        "description": "Error response. Most handlers return `{ \"detail\": \"...\" }`; some compatibility routes may return a plain text body when handled by net/http helpers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "additionalProperties": true,
        "example": {
          "name": "example",
          "status": "active",
          "settings": {}
        }
      },
      "AnyValue": {
        "nullable": true,
        "example": {
          "id": 1,
          "status": "ok"
        },
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}